requirements.txt
python-telegram-bot==13.15
python-docx==0.8.11
langdetect==1.0.9
TTS==0.14.3
pydub==0.25.1
ffmpeg-python==0.2.0
SQLAlchemy==2.0.19
aiohttp==3.8.5
asyncio==3.4.3
python-decouple==3.8
pyttsx3==2.90
gtts==2.3.2